+++ R/build_models.R
@@ -38,7 +38,11 @@ build_model <- function(X, y, method, param) {
     args <- args[!(names(args) %in% names(param))]
     args <- c(args, param)
     model <- do.call(ranger::ranger, args = args)
-  } 
+  } else if (method == "xgboost") { # extreme gradient boosting
+    args <- list(x = as.matrix(X), y = y)
+    args <- c(args, param)
+    model <- do.call(xgboost::xgboost, args = args)
+  }
   model
 }